icons: fix icons which have subtle compresion problems
These icons appear to have been compressed with a broken tool that results in an
IDAT that claims it needs a 512 byte window, but reaches backwards 1500 bytes in
the stream.  libpng < 1.6 is fine with this as zlib can still reach that data,
but libpng 1.6 requests strict decompression which turns this into an error.

These images have been recompressed and now load with libpng 1.6.